Implement end to end health checks
When a container is started on ECS, it will do a check first to see if
it can complete the DORA flow using perfDHCP. This test runs on the same
docker container as the server and will prevent the task from running if
the test was not successful.

This will prevent bad configuration taking down the entire cluster when
we implement zero downtime deployments.