Use spec helper to load all testing dependencies in one place
Previously we were relying on a single spec loading libraries for other
specs. This meant that the test suite only worked when we ran it as a whole
and it ran in the same order every time.

This change loads libraries used in testing to a spec_helper which is loaded
before a spec is run.

This change also updates the fixture path using a RSPEC_ROOT path
so that the test suite is not dependant on the current path