Remove container local acceptance tests
The introduction of the HA KEA module means that these tests will fail
depending on what state the peer server is in. If the standby is
currently serving traffic, the primary won't respond to the local
perfdhcp tests, meaning that primary will never come up.

We are also now pulling in KEA configurations from S3 and not rebuilding
the container on each change in the admin. This means that the tests will
only run on the first container boot.