Zero downtime deployments for KEA HA
After the container has been pushed to ecr, this script will start a
zero downtime deployment of the following services:

1. KEA Standby
2. KEA API
3. KEA Primary

It will double the task count and then scale it back in gracefully so
the new servers can take over traffic.

Tested with perfdhcp and this doesn't cause any downtime or
complications with peer heartbeats or lease syncs.