Move metrics loop out of agent into script
Currently if the metrics agent runs into an error, it will kill the ruby
process. This means that metrics won't be published.

KEA on first call sometimes returns an empty response when asking for
metrics, this can cause an error on first run.

By moving the loop to the script, the ruby process won't be killed on
error. This error still needs to be addressed.